+++ kernel/base/kmalloc.h
@@ -19,23 +19,23 @@
 #define MEM_MAGIC_NUM       0xABEEABEE
 
 typedef struct {
-    size_t magic;
-    size_t checkno;
-    size_t numpages;
-    size_t size;
-    char   filename[512];
-    size_t lineno;
+    uint64_t magic;
+    uint64_t checkno;
+    uint64_t numpages;
+    uint64_t size;
+    char     filename[512];
+    uint64_t lineno;
 } memory_metadata_t;
 
-extern size_t kmalloc_checkno;
+extern uint64_t kmalloc_checkno;
 
-void* kmalloc_core(uint64_t size, const char *func, size_t line);
-void kmfree_core(void* addr, const char *func, size_t line);
-void* kmrealloc_core(void* addr, size_t newsize, const char *func, size_t line);
+void* kmalloc_core(uint64_t size, const char *func, uint64_t line);
+void kmfree_core(void* addr, const char *func, uint64_t line);
+void* kmrealloc_core(void* addr, uint64_t newsize, const char *func, uint64_t line);
 
-void* kmalloc_chunk(uint64_t size, const char *func, size_t line);
-void kmfree_chunk(void* addr, const char *func, size_t line);
-void* kmrealloc_chunk(void* addr, size_t newsize, const char *func, size_t line);
+void* kmalloc_chunk(uint64_t size, const char *func, uint64_t line);
+void kmfree_chunk(void* addr, const char *func, uint64_t line);
+void* kmrealloc_chunk(void* addr, uint64_t newsize, const char *func, uint64_t line);
 
 #define kmalloc(x)          kmalloc_core(x, __func__, __LINE__)
 #define kmfree(x)           kmfree_core(x, __func__, __LINE__)
