+++ kernel/fs/pipefs.h
@@ -19,8 +19,8 @@ vfs_inode_t* pipefs_mount(vfs_inode_t *at);
 int64_t pipefs_mknode(vfs_tnode_t *this);
 int64_t pipefs_rmnode(vfs_tnode_t *this);
 vfs_tnode_t* pipefs_open(vfs_inode_t *this, const char *path);
-int64_t pipefs_read(vfs_inode_t *this, size_t offset, size_t len, void *buff);
-int64_t pipefs_write(vfs_inode_t *this, size_t offset, size_t len, const void *buff);
+int64_t pipefs_read(vfs_inode_t *this, uint64_t offset, uint64_t len, void *buff);
+int64_t pipefs_write(vfs_inode_t *this, uint64_t offset, uint64_t len, const void *buff);
 
 void pipefs_init(void);
