+++ kernel/fs/ramfs.h
@@ -42,9 +42,9 @@ vfs_inode_t *ramfs_mount(vfs_inode_t *at);
 int64_t ramfs_mknode(vfs_tnode_t *this);
 int64_t ramfs_rmnode(vfs_tnode_t *this);
 vfs_tnode_t *ramfs_open(vfs_inode_t *this, const char *path);
-int64_t ramfs_read(vfs_inode_t *this, size_t offset, size_t len, void *buff);
-int64_t ramfs_getdent(vfs_inode_t *this, size_t pos, vfs_dirent_t *dirent);
-int64_t ramfs_write(vfs_inode_t *this, size_t offset, size_t len, const void *buff);
+int64_t ramfs_read(vfs_inode_t *this, uint64_t offset, uint64_t len, void *buff);
+int64_t ramfs_getdent(vfs_inode_t *this, uint64_t pos, vfs_dirent_t *dirent);
+int64_t ramfs_write(vfs_inode_t *this, uint64_t offset, uint64_t len, const void *buff);
 int64_t ramfs_sync(vfs_inode_t *this);
 int64_t ramfs_refresh(vfs_inode_t *this);
