+++ kernel/mm/slab.h
@@ -7,22 +7,22 @@
 typedef struct slab_t {
     struct slab_t   *next;
     struct slab_t   *prev;
-    size_t          used;
+    uint64_t        used;
     void            **free;
     void            *base;
 } slab_t;
 
 typedef struct scache_t {
-    lock_t lock;
-    void   (*ctor)(struct scache_t *cache, void *obj);
-    void   (*dtor)(struct scache_t *cache, void *obj);
-    slab_t *full;
-    slab_t *partial;
-    slab_t *empty;
-    size_t size;
-    size_t truesize;
-    size_t alignment;
-    size_t slabobjcount;
+    lock_t   lock;
+    void     (*ctor)(struct scache_t *cache, void *obj);
+    void     (*dtor)(struct scache_t *cache, void *obj);
+    slab_t   *full;
+    slab_t   *partial;
+    slab_t   *empty;
+    uint64_t size;
+    uint64_t truesize;
+    uint64_t alignment;
+    uint64_t slabobjcount;
 } scache_t;
 
 void *slab_allocate(scache_t *cache);
@@ -30,7 +30,7 @@ void *slab_allocate(scache_t *cache);
 void slab_free(scache_t *cache, void *addr);
 
 scache_t *slab_newcache(
-    size_t size, size_t alignment,
+    uint64_t size, uint64_t alignment,
     void (*ctor)(scache_t *, void *),
     void (*dtor)(scache_t *, void *));
